A bit of magic with Hot Reload for Electron.
