+++ .gitignore
@@ -24,4 +24,4 @@ yarn-debug.log*
 yarn-error.log*
 
 # Electron
-electron/main.js
+src/electron/main.js
