+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "homepage": "./",
-  "main": "electron/main.js",
+  "main": "src/electron/main.js",
   "scripts": {
     "build": "react-scripts build",
     "eject": "react-scripts eject",
@@ -11,14 +11,15 @@
     "lint:fix": "eslint ./src --ext \".js,.jsx,.ts,.tsx\" --fix && stylelint \"./src/**/*.css\" --fix",
     "start": "nf start -p 3000",
     "test": "react-scripts test",
-    "electron": "tsc electron/main.ts && electron .",
-    "electron-start": "ts-node --skip-project electron/connect-electron.ts",
-    "react-start": "cross-env BROWSER=none react-scripts start"
+    "start-electron": "ts-node --skip-project src/electron/connect-electron.ts",
+    "watch-electron": "tsc --watch src/electron/main.ts",
+    "start-react": "cross-env BROWSER=none react-scripts start"
   },
   "dependencies": {
     "@material-ui/core": "^4.10.0",
     "@material-ui/icons": "^4.9.1",
     "classnames": "^2.2.6",
+    "electron-reload": "^1.5.0",
     "localforage": "^1.7.3",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
@@ -31,7 +32,7 @@
     "@types/classnames": "^2.2.10",
     "@types/jest": "^25.2.3",
     "@types/material-ui": "^0.21.7",
-    "@types/node": "14.0.5",
+    "@types/node": "^12.12.42",
     "@types/react": "^16.9.35",
     "@types/react-dom": "^16.9.8",
     "@types/react-router-dom": "^5.1.5",
@@ -40,6 +41,7 @@
     "cross-env": "^7.0.2",
     "electron": "^9.0.0",
     "electron-builder": "^22.7.0",
+    "electron-reloader": "^1.0.1",
     "eslint-plugin-react": "^7.20.0",
     "foreman": "^3.0.1",
     "stylelint": "^13.5.0",
