+++ src/electron/main.ts
@@ -1,11 +1,19 @@
 import { app, BrowserWindow } from 'electron';
+import * as reload from 'electron-reloader';
 import * as path from 'path';
 import * as url from 'url';
+
+// Prepare live reload
+try {
+  reload(module);
+} catch (err) {
+  console.log('Error encountered during hot reload!', err);
+}
 
 let mainWindow: BrowserWindow | null;
 
 function createWindow() {
-  mainWindow = new BrowserWindow({ width: 1024, height: 768 });
+  mainWindow = new BrowserWindow({ width: 1280, height: 1024, x: 0, y: 0 });
 
   mainWindow.loadURL(
     process.env.ELECTRON_START_URL ||
@@ -18,7 +26,7 @@ function createWindow() {
 
   mainWindow.on('closed', () => {
     mainWindow = null
-  })
+  });
 }
 
 app.on('ready', createWindow);
