+++ src/types/electron-reloader.d.ts
@@ -1,0 +1,1 @@
+declare module 'electron-reloader';
